+++ packages/core/src/repowise/core/ingestion/queries/c.scm
@@ -22,6 +22,37 @@
   name: (type_identifier) @symbol.name
 ) @symbol.def
 
+; typedef struct { ... } MyType;
+(type_definition
+  type: (struct_specifier)
+  declarator: (type_identifier) @symbol.name
+) @symbol.def
+
+; typedef enum { ... } MyEnum;
+(type_definition
+  type: (enum_specifier)
+  declarator: (type_identifier) @symbol.name
+) @symbol.def
+
+; #define MACRO_NAME ...
+(preproc_def
+  name: (identifier) @symbol.name
+) @symbol.def
+
+; #define FUNC_MACRO(x) ...
+(preproc_function_def
+  name: (identifier) @symbol.name
+  parameters: (preproc_params) @symbol.params
+) @symbol.def
+
+; Forward declarations: void func(int x);
+(declaration
+  declarator: (function_declarator
+    declarator: (identifier) @symbol.name
+    parameters: (parameter_list) @symbol.params
+  )
+) @symbol.def
+
 ; ---------------------------------------------------------------------------
 ; Imports (#include directives)
 ; ---------------------------------------------------------------------------